+++ pysnow/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from pysnow.client import *
+from pysnow.query import *
+from pysnow.request import *
+from pysnow.exceptions import *
+
+__author__ = "Robert Wikman <rbw@vault13.org>"
+__version__ = "0.4.0"
